refactor: TokenProvider 계층 변경
- domain 계층 보다는 util에서 전역적으로 주입 받는것이 맞다고 판단.
